fixed automod

+++ src/Structures/BotUtils.js
@@ -287,19 +287,18 @@ module.exports = class Tools {
 		return embed;
 	}
 	async doAction(message, action, user, reason, guildData) {
-		switch (action) {
-		case 'kick':
+		if(action == 'kick') {
 			user.kick({ reason: reason });
-			break;
-		case 'ban':
-			user.ban({ reason: reason });
-			break;
-		case 'mute':
-			message.member.roles.cache.add(guildData.roles.muterole);
-			break;
-		case 'delete':
-			message.delete();
-			break;
 		}
+		if(action == 'ban') {
+			user.ban({ reason: reason });
+		}
+		if(action == 'mute') {
+			message.member.roles.cache.add(guildData.roles.muterole);
+		}
+		if(action == 'delete') {
+			message.delete();
+		}
+
 	}
 };
